.claude/skills/fastapi-builder/assets/production-ready/requirements.txt
fastapi[standard]==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
pydantic[email]==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
pytest==8.3.4
httpx==0.28.1
